split nav bar into components
